Update current models structure
