+++ django_project/base/forms/csv_upload.py
@@ -3,7 +3,7 @@
 """
 
 from django import forms
-from fish.models.csv_document import CSVDocument
+from base.models.csv_document import CSVDocument
 
 
 class CsvUploadForm(forms.ModelForm):
