+++ django_project/base/models/__init__.py
@@ -1,3 +1,6 @@
 from base.models.location_type import *  # noqa
 from base.models.location_site import *  # noqa
+from base.models.csv_document import *  # noqa
+from base.models.iucn_status import *  # noqa
+from base.models.taxon import *  # noqa
 from base.models.profile import Profile
