+++ django_project/base/models/iucn_status.py
@@ -37,7 +37,7 @@ class IUCNStatus(models.Model):
     # noinspection PyClassicStyleClass
     class Meta:
         """Meta class for project."""
-        app_label = 'fish'
+        app_label = 'base'
         verbose_name_plural = 'IUCN Status'
         verbose_name = 'IUCN Status'
