+++ django_project/base/templates/csv_uploader.html
@@ -8,7 +8,7 @@
 {% block head %}
 
     <!-- Custom styles for this template -->
-    <link href="{% static "csv_uploader.css" %}" rel="stylesheet">
+    <link href="{% static "css/csv_uploader.css" %}" rel="stylesheet">
 
 {% endblock %}
