+++ django_project/base/templates/landing_page.html
@@ -82,7 +82,7 @@
 {% block foot %}
 
     <!-- Custom scripts for this template -->
-    <script src="{% static "js/creative.min.js" %}"></script>
+    <script src="{% static "js/creative.js" %}"></script>
 
     <!-- Plugin JavaScript -->
     <script src="{% static "libs/jquery-easing/jquery.easing.min.js" %}"></script>
