+++ django_project/base/tests/test_api_views.py
@@ -5,11 +5,13 @@ from fish.tests.model_factories import (
 )
 from base.tests.model_factories import (
     LocationSiteF,
+    TaxonF,
 )
 from base.api_views.location_site import (
     LocationSiteList,
     LocationSiteDetail,
 )
+from base.api_views.taxon import TaxonDetail
 
 
 class TestApiView(TestCase):
@@ -44,4 +46,18 @@ class TestApiView(TestCase):
         response = view(request, pk)
         self.assertTrue(
             len(response.data['fish_collection_records']) > 1
+        )
+
+    def test_get_taxon_by_id(self):
+        pk = 1
+        taxon = TaxonF.create(
+            pk=1,
+            common_name=u'Golden fish',
+        )
+        view = TaxonDetail.as_view()
+        request = self.factory.get('/api/taxon/' + str(pk))
+        response = view(request, str(pk))
+        self.assertEqual(
+            taxon.common_name,
+            response.data['common_name']
         )
