+++ django_project/fish/models/__init__.py
@@ -1,4 +1,1 @@
-from fish.models.iucn_status import *  # noqa
-from fish.models.taxon import *  # noqa
-from fish.models.csv_document import CSVDocument  # noqa
 from fish.models.fish_collection_record import *  # noqa
