+++ django_project/fish/tests/model_factories.py
@@ -5,37 +5,9 @@ from django.db.models import signals
 
 from fish.models import (
     FishCollectionRecord,
-    IUCNStatus,
-    Taxon,
 )
-from base.tests.model_factories import LocationSiteF
+from base.tests.model_factories import LocationSiteF, TaxonF
 from core.tests.model_factories import UserF
-
-
-class IUCNStatusF(factory.django.DjangoModelFactory):
-    """
-    Iucn status factory
-    """
-    class Meta:
-        model = IUCNStatus
-
-    category = factory.Sequence(lambda n: u'Test name %s' % n)
-    sensitive = False
-
-
-@factory.django.mute_signals(signals.pre_save)
-class TaxonF(factory.django.DjangoModelFactory):
-    """
-    Taxon factory
-    """
-    class Meta:
-        model = Taxon
-
-    iucn_status = factory.SubFactory(IUCNStatusF)
-    common_name = factory.Sequence(lambda n: u'Test common name %s' % n)
-    scientific_name = factory.Sequence(
-            lambda n: u'Test scientific name %s' % n)
-    author = factory.Sequence(lambda n: u'Test author %s' % n)
 
 
 @factory.django.mute_signals(signals.post_save)
