+++ django_project/fish/tests/test_api_views.py
@@ -2,13 +2,11 @@ from django.test import TestCase
 from rest_framework.test import APIRequestFactory
 from fish.tests.model_factories import (
     FishCollectionRecordF,
-    TaxonF,
 )
 from fish.api_views.fish_collection_record import (
     FishCollectionList,
     FishCollectionDetail
 )
-from fish.api_views.taxon import TaxonDetail
 
 
 class TestApiView(TestCase):
@@ -39,18 +37,4 @@ class TestApiView(TestCase):
         self.assertEqual(
             self.fish_collection_1.original_species_name,
             response.data['original_species_name']
-        )
-
-    def test_get_taxon_by_id(self):
-        pk = 1
-        taxon = TaxonF.create(
-            pk=1,
-            common_name=u'Golden fish',
-        )
-        view = TaxonDetail.as_view()
-        request = self.factory.get('/api/taxon/' + str(pk))
-        response = view(request, str(pk))
-        self.assertEqual(
-            taxon.common_name,
-            response.data['common_name']
         )
